+++ writer/linker.cxx
@@ -185,7 +185,7 @@ void Linker::write(Writer& writer, bool doShuffle)
 		for (const auto& entry : mMap)
 		{
 			printf("0x%06x 0x%06x 0x%06x 0x%06x %s  %s %s\n",
-				entry.begin, entry.end, entry.end - entry.begin,
+				(u32)entry.begin, (u32)entry.end, (u32)entry.end - (u32)entry.begin,
 				entry.restrict.alignment,
 				entry.restrict.options & LinkingRestriction::Static ? "true " : "false",
 				entry.restrict.options & LinkingRestriction::Leaf ? "true " : "false",
